requirements[1].txt
tensorflow
numpy
opencv-python
Pillow
matplotlib
